+++ src/main/java/umc/meme/auth/domain/artist/entity/Artist.java
@@ -6,10 +6,10 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.SuperBuilder;
-import umc.meme.auth.domain.artist.entity.enums.Category;
-import umc.meme.auth.domain.artist.entity.enums.MakeupLocation;
-import umc.meme.auth.domain.artist.entity.enums.Region;
-import umc.meme.auth.domain.artist.entity.enums.WorkExperience;
+import umc.meme.auth.global.enums.Category;
+import umc.meme.auth.global.enums.MakeupLocation;
+import umc.meme.auth.global.enums.Region;
+import umc.meme.auth.global.enums.WorkExperience;
 import umc.meme.auth.domain.user.entity.User;
 import umc.meme.auth.domain.user.dto.UserRequest;
 import umc.meme.auth.global.enums.DayOfWeek;
