--- src/main/java/umc/meme/auth/domain/artist/entity/enums/Gender.java
@@ -1,5 +1,0 @@
-package umc.meme.auth.domain.artist.entity.enums;
-
-public enum Gender {
-    MALE, FEMALE
-}
