+++ src/main/java/umc/meme/auth/domain/model/entity/Model.java
@@ -5,8 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
-import umc.meme.auth.domain.model.entity.enums.PersonalColor;
-import umc.meme.auth.domain.model.entity.enums.SkinType;
+import umc.meme.auth.global.enums.PersonalColor;
+import umc.meme.auth.global.enums.SkinType;
 import umc.meme.auth.domain.user.entity.User;
 
 @SuperBuilder @Getter
