+++ src/main/java/umc/meme/auth/domain/user/dto/UserRequest.java
@@ -4,12 +4,9 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import umc.meme.auth.domain.artist.entity.enums.*;
-import umc.meme.auth.domain.model.entity.enums.PersonalColor;
-import umc.meme.auth.domain.model.entity.enums.SkinType;
-import umc.meme.auth.global.enums.DayOfWeek;
-import umc.meme.auth.global.enums.Provider;
-import umc.meme.auth.global.enums.Times;
+import umc.meme.auth.global.enums.PersonalColor;
+import umc.meme.auth.global.enums.SkinType;
+import umc.meme.auth.global.enums.*;
 
 import java.util.List;
 import java.util.Map;
