+++ src/main/java/umc/meme/auth/global/enums/Gender.java
@@ -1,0 +1,5 @@
+package umc.meme.auth.global.enums;
+
+public enum Gender {
+    MALE, FEMALE
+}
